+++ src/index.ts
@@ -12,7 +12,7 @@ export { enumType } from "./definitions/enumType";
 export { extendType } from "./definitions/extendType";
 export { inputObjectType } from "./definitions/inputObjectType";
 export { interfaceType } from "./definitions/interfaceType";
-export { objectType } from "./definitions/objectType";
+export { objectType, queryType, mutationType } from "./definitions/objectType";
 export { scalarType } from "./definitions/scalarType";
 export { unionType } from "./definitions/unionType";
 export { nexusWrappedType } from "./definitions/wrapping";